Extract a module inside chip for code specific to the ESP32 and variants
